+++ NEWS.md
@@ -1,3 +1,10 @@
-# criticalpath 0.1.0
+# criticalpath 0.1.0.01
+
+Fixed some bugs:
+
+- In gantt_matrix function there was a bug when activity's duration is zero.
+- In xy_gantt_matrix function a "+ 1" was missing.
+
 
 First version published.
+